+++ src/applications/administration-users-app/users/edit-user/edit-user.component.ts
@@ -340,7 +340,7 @@ export class EditUserComponent implements OnInit, OnDestroy {
     if (!event) {
       this._selectedRoleDescription = this._roles[0] ? this._roles[0].description : '';
     } else {
-      const relevantRole = this._roles.find(role => event === role.id.toString() || event.value === role.id);
+      const relevantRole = this._roles.find(role => event === role.id.toString() || event.value === role.id.toString());
       if (relevantRole) {
         this._selectedRoleDescription = relevantRole.description;
       }