+++ src/applications/content-playlists-app/playlist/playlist-content/rule-based/playlist-rule/playlist-entries-data-provider.service.ts
@@ -306,6 +306,7 @@ export class PlaylistEntriesDataProvider implements EntriesDataProvider, OnDestr
       sortBy: 'plays',
       sortDirection: SortDirection.Desc,
       createdAt: { fromDate: null, toDate: null },
+      lastPlayedAt: null,
       scheduledAt: { fromDate: null, toDate: null },
       mediaTypes: [],
       timeScheduling: [],