+++ src/shared/content-shared/entries/entries-store/entries-store-data-provider.service.ts
@@ -119,6 +119,11 @@ export class EntriesStoreDataProvider implements EntriesDataProvider, OnDestroy 
             }
           }
 
+          // filter 'lastPlayedAt'
+          if (data.lastPlayedAt) {
+              filter.lastPlayedAtLessThanOrEqual = KalturaUtils.getEndDateValue(new Date(data.lastPlayedAt * 1000));
+          }
+
           // filters of joined list
           this._updateFilterWithJoinedList(data.mediaTypes, filter, 'mediaTypeIn');
           this._updateFilterWithJoinedList(data.ingestionStatuses, filter, 'statusIn');
@@ -374,6 +379,7 @@ export class EntriesStoreDataProvider implements EntriesDataProvider, OnDestroy 
       sortBy: 'createdAt',
       sortDirection: SortDirection.Desc,
       createdAt: { fromDate: null, toDate: null },
+      lastPlayedAt: null,
       scheduledAt: { fromDate: null, toDate: null },
       mediaTypes: [],
       timeScheduling: [],